repairing some pivotal table issues
The pivotal table query didn't work if the local key was different than the primary key. Now it should work only if the name of the fields in the pivotal table have the following format: foreign_table_foreign_key, another_foreign_table_another_key. For example, on issue #11:

If you have a programs table and a users table and want to join them with the help of a pivotal table with programs->id and users->user_id (and not users->id), you should create the following fields inside the "programs_users" table: "program_id" and "user_user_id" (I sure hope that makes sense, @papakey ).