Order by relationship issue
Fix undefined index in order_by inside relationship #171. 

When the relationship is empty set it to NULL to prevent object_to_array conversion error. 
Row with empty relationship will be place at the bottom of the list regardless of the direction (ASC or DESC) 
This can be override to place empty relationship at the top  by appending  _NULL_TOP to the direction (asc_null_top or desc_null_top)

I have tested with a limited set of data. This should be tested with a larger data set